2021 organizing documents (in PDF, ODT, DOCX
The INVITATION is what attendees responded to. Ultimately six creators were able to commit to the requirements mentioned in the invitation. We had an initial organizing meeting and then two Zooms (each focusing on feedback for three works). People could provide additional feedback after the Zooms. 
The IDEAS FOR FUTURE SESSIONS is the paraphrased feedback from participants about how they felt the processes could be improved in the future.